+++ vite.config.ts
@@ -23,11 +23,10 @@ export default defineConfig({
       }
     },
     rollupOptions: {
-      external: ['@mediapipe/tasks-vision', '@tensorflow/tfjs', 'react'],
+      external: ['@mediapipe/tasks-vision', 'react'],
       output: {
         globals: {
           '@mediapipe/tasks-vision': 'MediaPipeTasksVision',
-          '@tensorflow/tfjs': 'tf',
           'react': 'React'
         }
       }